Initial pass at implementing dynamic type checking for the cython integrator.
Implements this via fused types on the `y0` input array buffer.
For convenience (and maybe necessity), I've removed a bunch of the older numpy buffer syntax.
Preliminary tests on toy code seem to indicate they are not needed if
a typed memory view assigned to the output of the numpy allocators is used anyway.

I've also recast all the RK constants to doubles and rely on cython to
correctly cast them to double complex when necessary.

Finally, I have not tested the interpolation part at all.
I strongly suspect there are runtime issues to be fixed there as well.